+++ fake-cucumber/javascript/test/TestStepTest.ts
@@ -63,6 +63,7 @@ describe('TestStep', () => {
       'some-test-step-id',
       {
         text: 'an undefined step',
+        type: messages.PickleStepType.UNKNOWN,
         astNodeIds: [],
         id: '1',
       },
@@ -77,6 +78,7 @@ describe('TestStep', () => {
       'some-test-step-id',
       {
         text: 'a failed step',
+        type: messages.PickleStepType.OUTCOME,
         astNodeIds: [],
         id: '1',
       },
@@ -91,6 +93,7 @@ describe('TestStep', () => {
       'some-test-step-id',
       {
         text: 'a failed step',
+        type: messages.PickleStepType.ACTION,
         astNodeIds: [],
         id: '1',
       },
@@ -105,6 +108,7 @@ describe('TestStep', () => {
       'some-test-step-id',
       {
         text: 'a passed step',
+        type: messages.PickleStepType.CONTEXT,
         astNodeIds: [],
         id: '1',
       },
@@ -119,6 +123,7 @@ describe('TestStep', () => {
       'some-test-step-id',
       {
         text: 'a pending step',
+        type: messages.PickleStepType.UNKNOWN,
         astNodeIds: [],
         id: '1',
       },
@@ -133,6 +138,7 @@ describe('TestStep', () => {
       'some-test-step-id',
       {
         text: 'a skipped step',
+        type: messages.PickleStepType.UNKNOWN,
         astNodeIds: [],
         id: '1',
       },