+++ gherkin/dotnet/Gherkin/StepKeywordType.cs
@@ -1,0 +1,12 @@
+﻿namespace Gherkin
+{
+    public enum StepKeywordType
+    {
+        Unspecified,
+        Context,
+        Action,
+        Outcome,
+        Conjunction,
+        Unknown
+    }
+}