+++ gherkin/java/src/main/java/io/cucumber/gherkin/GherkinDialect.java
@@ -1,95 +1,137 @@
 package io.cucumber.gherkin;
 
-import com.eclipsesource.json.JsonArray;
-import com.eclipsesource.json.JsonObject;
-import com.eclipsesource.json.JsonValue;
+import io.cucumber.messages.types.StepKeywordType;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
-import static java.util.Objects.requireNonNull;
+import static java.util.Collections.unmodifiableList;
 
 public final class GherkinDialect {
-    private final JsonObject keywords;
     private final String language;
+    private final String name;
+    private final String nativeName;
+    private final List<String> featureKeywords;
+    private final List<String> ruleKeywords;
+    private final List<String> scenarioKeywords;
+    private final List<String> scenarioOutlineKeywords;
+    private final List<String> backgroundKeywords;
+    private final List<String> examplesKeywords;
+    private final List<String> givenKeywords;
+    private final List<String> whenKeywords;
+    private final List<String> thenKeywords;
+    private final List<String> andKeywords;
+    private final List<String> butKeywords;
+    private final List<String> stepKeywords;
+    private final Map<String, List<StepKeywordType>> stepKeywordsTypes;
 
-    public GherkinDialect(String language, JsonObject keywords) {
-        this.language = requireNonNull(language);
-        this.keywords = requireNonNull(keywords);
+    GherkinDialect(String language, String name, String nativeName, List<String> featureKeywords, List<String> ruleKeywords, List<String> scenarioKeywords, List<String> scenarioOutlineKeywords, List<String> backgroundKeywords, List<String> examplesKeywords, List<String> givenKeywords, List<String> whenKeywords, List<String> thenKeywords, List<String> andKeywords, List<String> butKeywords) {
+        this.language = language;
+        this.name = name;
+        this.nativeName = nativeName;
+        this.featureKeywords = featureKeywords;
+        this.ruleKeywords = ruleKeywords;
+        this.scenarioKeywords = scenarioKeywords;
+        this.scenarioOutlineKeywords = scenarioOutlineKeywords;
+        this.backgroundKeywords = backgroundKeywords;
+        this.examplesKeywords = examplesKeywords;
+        this.givenKeywords = givenKeywords;
+        this.whenKeywords = whenKeywords;
+        this.thenKeywords = thenKeywords;
+        this.andKeywords = andKeywords;
+        this.butKeywords = butKeywords;
+
+        List<String> stepKeywords = new ArrayList<>();
+        stepKeywords.addAll(givenKeywords);
+        stepKeywords.addAll(whenKeywords);
+        stepKeywords.addAll(thenKeywords);
+        stepKeywords.addAll(andKeywords);
+        stepKeywords.addAll(butKeywords);
+        this.stepKeywords = unmodifiableList(stepKeywords);
+
+        Map<String, List<StepKeywordType>> stepKeywordsTypes = new HashMap<>();
+        addStepKeywordsTypes(stepKeywordsTypes, getGivenKeywords(), StepKeywordType.CONTEXT);
+        addStepKeywordsTypes(stepKeywordsTypes, getWhenKeywords(), StepKeywordType.ACTION);
+        addStepKeywordsTypes(stepKeywordsTypes, getThenKeywords(), StepKeywordType.OUTCOME);
+
+        List<String> conjunctionKeywords = new ArrayList<>();
+        conjunctionKeywords.addAll(getAndKeywords());
+        conjunctionKeywords.addAll(getButKeywords());
+        addStepKeywordsTypes(stepKeywordsTypes, conjunctionKeywords, StepKeywordType.CONJUNCTION);
+        this.stepKeywordsTypes = stepKeywordsTypes;
+    }
+
+    private static void addStepKeywordsTypes(Map<String, List<StepKeywordType>> h, List<String> keywords, StepKeywordType type) {
+        for (String keyword : keywords) {
+            if (!h.containsKey(keyword))
+                h.put(keyword, new ArrayList<>());
+            h.get(keyword).add(type);
+        }
     }
 
     public List<String> getFeatureKeywords() {
-        return toStringList(this.keywords.get("feature").asArray());
-    }
-
-    private static List<String> toStringList(JsonArray array) {
-        List<String> result = new ArrayList<>();
-        for (JsonValue jsonValue : array) {
-            result.add(jsonValue.asString());
-        }
-        return result;
+        return featureKeywords;
     }
 
     public String getName() {
-        return keywords.getString("name", null);
+        return name;
     }
 
     public String getNativeName() {
-        return keywords.getString("native", null);
+        return nativeName;
     }
 
     public List<String> getRuleKeywords() {
-        return toStringList(keywords.get("rule").asArray());
+        return ruleKeywords;
     }
 
     public List<String> getScenarioKeywords() {
-        return toStringList(keywords.get("scenario").asArray());
+        return scenarioKeywords;
     }
 
     public List<String> getScenarioOutlineKeywords() {
-        return toStringList(keywords.get("scenarioOutline").asArray());
+        return scenarioOutlineKeywords;
     }
 
     public List<String> getStepKeywords() {
-        List<String> result = new ArrayList<>();
-        result.addAll(getGivenKeywords());
-        result.addAll(getWhenKeywords());
-        result.addAll(getThenKeywords());
-        result.addAll(getAndKeywords());
-        result.addAll(getButKeywords());
-        return result;
+        return stepKeywords;
+    }
+
+    public List<StepKeywordType> getStepKeywordTypes(String keyword) {
+        return stepKeywordsTypes.get(keyword);
     }
 
     public List<String> getBackgroundKeywords() {
-        return toStringList(keywords.get("background").asArray());
+        return backgroundKeywords;
     }
 
     public List<String> getExamplesKeywords() {
-        return toStringList(keywords.get("examples").asArray());
+        return examplesKeywords;
     }
 
     public List<String> getGivenKeywords() {
-        return toStringList(keywords.get("given").asArray());
+        return givenKeywords;
     }
 
     public List<String> getWhenKeywords() {
-        return toStringList(keywords.get("when").asArray());
+        return whenKeywords;
     }
 
     public List<String> getThenKeywords() {
-        return toStringList(keywords.get("then").asArray());
+        return thenKeywords;
     }
 
     public List<String> getAndKeywords() {
-        return toStringList(keywords.get("and").asArray());
+        return andKeywords;
     }
 
     public List<String> getButKeywords() {
-        return toStringList(keywords.get("but").asArray());
+        return butKeywords;
     }
 
     public String getLanguage() {
         return language;
     }
 }
-