+++ gherkin/php/src/TokenMatch.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Cucumber\Gherkin;
 
 use Cucumber\Gherkin\Parser\TokenType;
+use Cucumber\Messages\Step\KeywordType;
 
 final class TokenMatch
 {
@@ -16,6 +17,7 @@ final class TokenMatch
         public readonly GherkinDialect $gherkinDialect,
         public readonly int $indent,
         public readonly string $keyword,
+        public readonly KeywordType|null $keywordType,
         public readonly string $text,
         public readonly array $items,
         public readonly Location $location,